+++ apps/meteor/app/livechat/client/ui.js
@@ -1,14 +1,4 @@
-import { settings } from '../../settings/client';
-import { hasAllPermission } from '../../authorization/client';
-import { AccountBox, MessageTypes } from '../../ui-utils/client';
-
-AccountBox.addItem({
-	name: 'Omnichannel',
-	icon: 'headset',
-	href: '/omnichannel/current',
-	sideNav: 'omnichannelFlex',
-	condition: () => settings.get('Livechat_enabled') && hasAllPermission('view-livechat-manager'),
-});
+import { MessageTypes } from '../../ui-utils/client';
 
 MessageTypes.registerType({
 	id: 'livechat-close',
