+++ apps/meteor/app/ui-utils/client/index.ts
@@ -1,6 +1,4 @@
 import './lib/messageActionDefault';
-
-export { AccountBox } from './lib/AccountBox';
 export { MessageAction } from './lib/MessageAction';
 export { messageBox } from './lib/messageBox';
 export { readMessage } from './lib/readMessages';
