+++ apps/meteor/client/sidebar/header/actions/hooks/useAdministrationItems.tsx
@@ -1,8 +1,16 @@
-import { useTranslation, useRoute, useMethod, useSetModal, useRole, useRouter } from '@rocket.chat/ui-contexts';
+import {
+	useTranslation,
+	useRoute,
+	useMethod,
+	useSetModal,
+	useRole,
+	useRouter,
+	useAtLeastOnePermission,
+	usePermission,
+} from '@rocket.chat/ui-contexts';
 import { useQuery } from '@tanstack/react-query';
 import React from 'react';
 
-import type { AccountBoxItem } from '../../../../../app/ui-utils/client/lib/AccountBox';
 import type { UpgradeTabVariant } from '../../../../../lib/upgradeTab';
 import { getUpgradeTabLabel, isFullyFeature } from '../../../../../lib/upgradeTab';
 import Emoji from '../../../../components/Emoji';
@@ -10,17 +18,45 @@ import type { GenericMenuItemProps } from '../../../../components/GenericMenu/Ge
 import RegisterWorkspaceModal from '../../../../views/admin/cloud/modals/RegisterWorkspaceModal';
 import { useUpgradeTabParams } from '../../../../views/hooks/useUpgradeTabParams';
 
-type useAdministrationItemProps = {
-	accountBoxItems: AccountBoxItem[];
-	showWorkspace: boolean;
-};
-export const useAdministrationItems = ({ accountBoxItems, showWorkspace }: useAdministrationItemProps): GenericMenuItemProps[] => {
+const ADMIN_PERMISSIONS = [
+	'view-statistics',
+	'run-import',
+	'view-user-administration',
+	'view-room-administration',
+	'create-invite-links',
+	'manage-cloud',
+	'view-logs',
+	'manage-sounds',
+	'view-federation-data',
+	'manage-email-inbox',
+	'manage-emoji',
+	'manage-outgoing-integrations',
+	'manage-own-outgoing-integrations',
+	'manage-incoming-integrations',
+	'manage-own-incoming-integrations',
+	'manage-oauth-apps',
+	'access-mailer',
+	'manage-user-status',
+	'access-permissions',
+	'access-setting-permissions',
+	'view-privileged-setting',
+	'edit-privileged-setting',
+	'manage-selected-settings',
+	'view-engagement-dashboard',
+	'view-moderation-console',
+];
+
+export const useAdministrationItems = (): GenericMenuItemProps[] => {
 	const router = useRouter();
 	const t = useTranslation();
 
+	const shouldShowAdminMenu = useAtLeastOnePermission(ADMIN_PERMISSIONS);
+
 	const { tabType, trialEndDate, isLoading } = useUpgradeTabParams();
 	const shouldShowEmoji = isFullyFeature(tabType);
+
 	const label = getUpgradeTabLabel(tabType);
+
 	const isAdmin = useRole('admin');
 	const setModal = useSetModal();
 
@@ -36,7 +72,17 @@ export const useAdministrationItems = ({ accountBoxItems, showWorkspace }: useAd
 	const adminRoute = useRoute('admin-index');
 	const upgradeRoute = useRoute('upgrade');
 	const cloudRoute = useRoute('cloud');
+
+	const omnichannel = usePermission('view-livechat-manager');
+
 	const showUpgradeItem = !isLoading && tabType;
+
+	const omnichannelItem: GenericMenuItemProps = {
+		id: 'omnichannel',
+		content: t('Omnichannel'),
+		icon: 'headset',
+		onClick: () => router.navigate('/omnichannel/current'),
+	};
 
 	const upgradeItem: GenericMenuItemProps = {
 		id: 'showUpgradeItem',
@@ -71,24 +117,10 @@ export const useAdministrationItems = ({ accountBoxItems, showWorkspace }: useAd
 		},
 	};
 
-	const accountBoxItem: GenericMenuItemProps[] = accountBoxItems.map((item, key) => {
-		const action = () => {
-			if (item.href) {
-				router.navigate(item.href);
-			}
-		};
-		return {
-			id: `account-box-item-${key}`,
-			content: t(item.name),
-			icon: item.icon,
-			onClick: action,
-		};
-	});
-
 	return [
-		...(showUpgradeItem ? [upgradeItem] : []),
-		...(isAdmin ? [adminItem] : []),
-		...(showWorkspace ? [workspaceItem] : []),
-		...(accountBoxItems.length ? accountBoxItem : []),
-	];
+		showUpgradeItem && upgradeItem,
+		isAdmin && adminItem,
+		omnichannel && omnichannelItem,
+		shouldShowAdminMenu && workspaceItem,
+	].filter(Boolean) as GenericMenuItemProps[];
 };
