+++ apps/meteor/client/sidebar/header/actions/hooks/useAuditItems.tsx
@@ -1,17 +1,22 @@
-import { useTranslation, useRoute } from '@rocket.chat/ui-contexts';
+import { useTranslation, useRoute, usePermission } from '@rocket.chat/ui-contexts';
 
+import { useHasLicenseModule } from '../../../../../ee/client/hooks/useHasLicenseModule';
 import type { GenericMenuItemProps } from '../../../../components/GenericMenu/GenericMenuItem';
 
-type useAuditItemsProps = {
-	showAudit: boolean;
-	showAuditLog: boolean;
-};
+export const useAuditItems = (): GenericMenuItemProps[] => {
+	const hasAuditLicense = useHasLicenseModule('auditing') === true;
 
-export const useAuditItems = ({ showAudit, showAuditLog }: useAuditItemsProps): GenericMenuItemProps[] => {
+	const hasAuditPermission = usePermission('can-audit') && hasAuditLicense;
+	const hasAuditLogPermission = usePermission('can-audit-log') && hasAuditLicense;
+
 	const t = useTranslation();
 
 	const auditHomeRoute = useRoute('audit-home');
 	const auditSettingsRoute = useRoute('audit-log');
+
+	if (!hasAuditPermission && !hasAuditLogPermission) {
+		return [];
+	}
 
 	const auditMessageItem: GenericMenuItemProps = {
 		id: 'messages',
@@ -26,5 +31,5 @@ export const useAuditItems = ({ showAudit, showAuditLog }: useAuditItemsProps): 
 		onClick: () => auditSettingsRoute.push(),
 	};
 
-	return [...(showAudit ? [auditMessageItem] : []), ...(showAuditLog ? [auditLogItem] : [])];
+	return [hasAuditPermission && auditMessageItem, hasAuditLogPermission && auditLogItem].filter(Boolean) as GenericMenuItemProps[];
 };
